requirements.txt
alembic==1.12.1
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flask-sqlalchemy==3.1.1
greenlet==3.0.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
werkzeug==3.0.1
zipp==3.17.0
